pyproject.toml
[tool.poetry]
name = "can_stethoscope"
version = "0.1.0"
description = "Troubleshooting tools used to detect key measurements in the analog values of a CAN bus."
license = "MIT"
authors = [
    "Emerson Maki <makix222@gmail.com>",
    "Jamison Jangula <jamisonjangula@gmail.com>",
    "Eric Pierce <epierce137@gmail.com>",
    "Patrick Mubiayi <patrick.mubs007@gmail.com>"]
readme = "README.md"
repository = "https://github.com/jamisonjangula/can_stethoscope"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.4"
matplotlib = "^3.4.3"
python-can = "^3.3.4"

[tool.poetry.scripts]
can_stetho = "can_stethoscope.cli:main"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"